+++ R/ensureDBTables.R
@@ -30,51 +30,58 @@ ensureDBTables = function(src, projRecv, deviceID, quiet = FALSE) {
   tables = dplyr::src_tbls(src)
   
   isRecvDB = is.character(projRecv)
-  
   if (! "meta" %in% tables) {
-    if (missing(projRecv))
-      stop("you must specify a project number or receiver serial number for a new database")
-    sql("
-create table meta (
-key  character not null unique primary key, -- name of key for meta data
-val  character                              -- character string giving meta data; might be in JSON format
-)
-");
-if (isRecvDB)  {
-  if (missing(deviceID) || ! isTRUE(is.numeric(deviceID))) {
-    stop("must specify deviceID for new receiver database")
-  }
-  if (grepl("^SG", projRecv)) {
-    type = "SENSORGNOME"
-    model = substring(projRecv, 8, 11)
-  } else {
-    type = "Lotek"
-    model = getLotekModel(projRecv)
-  }
-  sql("
-insert into meta (key, val)
-values
-('dbType', 'receiver'),
-('recvSerno', '%s'),
-('recvType', '%s'),
-('recvModel', '%s'),
-('deviceID', '%d')
-",
-      projRecv,
-      type,
-      model,
-      as.integer(deviceID))
-} else if (is.numeric(projRecv)) {
-  sql("
-insert into meta (key, val)
-values
-('dbType', 'tag'),
-('tagProject', %d)
-",
-      projRecv)
-} else {
-  stop("projRecv must be an integer motus project ID or a character receiver serial number")
-}
+    if (missing(projRecv)) stop("you must specify a project number or receiver serial number for a new database")
+    sql(paste("create table meta (",
+               "key  character not null unique primary key, -- name of key for meta data",
+               "val  character                              -- character string giving meta data; might be in JSON format)",
+               ");", sep = "\n"))
+    if (isRecvDB)  {
+      if (missing(deviceID) || ! isTRUE(is.numeric(deviceID))) {
+        stop("must specify deviceID for new receiver database")
+      }
+      type <- tolower(stringr::str_remove(projRecv, "-(.)*$"))
+      
+      if (type == "sg") {
+        type <- "SENSORGNOME"
+        model <- substring(projRecv, 8, 11)
+      } else if(type == "lotek") {
+        type <- "Lotek"
+        model <- getLotekModel(projRecv)
+      } else if(type == "ctt") {
+        type <- "CTT"
+        n <- nchar(projRecv)
+        if(n == 15 + 4) {
+          model <- "V1"
+        } else if(n == 12 + 4) {
+          model <- "V2"
+        } else stop("Unexpected model for CTT receivers: ", projRecv, call. = FALSE)
+      } else {
+        stop("Unexpected receiver type: ", type, call. = FALSE)
+      }
+        
+      sql(paste("insert into meta (key, val)",
+                 "values",
+                 "('dbType', 'receiver'),",
+                 "('recvSerno', '%s'),",
+                 "('recvType', '%s'),",
+                 "('recvModel', '%s'),",
+                 "('deviceID', '%d')", sep = "\n"),
+          projRecv,
+          type,
+          model,
+          as.integer(deviceID))
+
+      
+    } else if (is.numeric(projRecv)) {
+      sql(paste("insert into meta (key, val)", 
+                 "values",
+                 "('dbType', 'tag'),",
+                 "('tagProject', %d)", sep = "\n"),
+          projRecv)
+    } else {
+      stop("projRecv must be an integer motus project ID or a character receiver serial number")
+    }
   }
   
   if (! "gps" %in% tables) {
@@ -516,7 +523,7 @@ CREATE INDEX IF NOT EXISTS runsFilters_filterID_runID_motusTagID ON runsFilters 
 
 makeTables <- function(type, name = type) {
   if(type == "tagAmbig") {
-    s <- paste0("CREATE TABLE ", name, " (
+    s <- paste("CREATE TABLE ", name, " (
     ambigID INTEGER PRIMARY KEY NOT NULL,  -- identifier of group of tags which are ambiguous (identical). Will be negative
     masterAmbigID INTEGER,                 -- master ID of this ambiguity group, once different receivers have been combined
     motusTagID1 INT,                       -- motus ID of tag in group.
@@ -528,7 +535,7 @@ makeTables <- function(type, name = type) {
     ambigProjectID INT                     -- negative ambiguity ID of deployment project. refers to key ambigProjectID in table projAmbig
 );")
   } else if(type == "nodeData") {
-  s <- paste0("CREATE TABLE IF NOT EXISTS ", name, " (",
+  s <- paste("CREATE TABLE IF NOT EXISTS ", name, " (",
     "nodeDataID BIGINT PRIMARY KEY NOT NULL,",
     "batchID INTEGER NOT NULL,",
     "ts FLOAT NOT NULL,",